todo antes del colapso (despues)

+++ src/main/java/com/example/fidelidad_api/repositories/MarcaRepositorio.java
@@ -3,5 +3,7 @@ package com.example.fidelidad_api.repositories;
 import com.example.fidelidad_api.models.Marca;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MarcaRepositorio extends JpaRepository<Marca, Integer> {
+public interface MarcaRepositorio extends JpaRepository<Marca, Long> {
+    // No se expone ningún método personalizado de escritura
 }
+
