+++ src/main/java/com/example/fidelidad_api/repositories/TipoIdentificacionRepositorio.java
@@ -3,6 +3,4 @@ package com.example.fidelidad_api.repositories;
 import com.example.fidelidad_api.models.TipoIdentificacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoIdentificacionRepositorio extends JpaRepository<TipoIdentificacion, Integer> {
-}
-
+public interface TipoIdentificacionRepositorio extends JpaRepository<TipoIdentificacion, Long> {}
